third_party/WebKit/LayoutTests/inspector/tracing/timeline-style-recalc-with-invalidations-expected.txt

Tests the Timeline API instrumentation of style recalc events with invalidations.

PASS
PASS
PASS


Running: testLocalFrame
first recalc style invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../inspector/tracing/timeline-style-recalc-with-invalidations.html:10}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementOne'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../inspector/tracing/timeline-style-recalc-with-invalidations.html:11}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementTwo'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]

Running: multipleStyleRecalcs
first recalc style invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../inspector/tracing/timeline-style-recalc-with-invalidations.html:25}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementOne'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
second recalc style invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../inspector/tracing/timeline-style-recalc-with-invalidations.html:27}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementOne'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../inspector/tracing/timeline-style-recalc-with-invalidations.html:28}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementTwo'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
third recalc style invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../inspector/tracing/timeline-style-recalc-with-invalidations.html:30}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementOne'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../inspector/tracing/timeline-style-recalc-with-invalidations.html:31}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementTwo'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../inspector/tracing/timeline-style-recalc-with-invalidations.html:32}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElementThree'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]

Running: testSubframe
first recalc style invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../inspector/tracing/timeline-style-recalc-with-invalidations.html:43}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "BODY"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../inspector/tracing/timeline-style-recalc-with-invalidations.html:44}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
